pyproject.toml
[project]
name = "vlsp-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argparse>=1.4.0",
    "boto3>=1.39.4",
    "google-genai>=1.25.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
